requirements.txt
fastapi
fastmcp
uvicorn
google-generativeai
python-dotenv
